requirements.txt
blinker==1.7.0
click==8.1.7
coloredlogs==15.0.1
Flask==3.0.3
flatbuffers==24.3.25
gunicorn==21.2.0
humanfriendly==10.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.3
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
sympy==1.12
Werkzeug==3.0.2